version=11.5.29  ellipsoids
# new features for ellipsoid rendering: 
#  
#  set ellipsoidAxes     [false]
#  set ellipsoidArcs     [false]
#  set ellipsoidFill     [false]
#  set ellipsoidDots     [true]
#  set ellipsoidDotCount [200]
#
#  [dots|stars|spacefill] ADPMIN [percent]
#  [dots|stars|spacefill] ADPMAX [percent]
#

+++ Jmol/src/org/jmol/shape/AxesRenderer.java
@@ -48,7 +48,7 @@ public class AxesRenderer extends FontLineShapeRenderer {
 
   protected void render() {
     Axes axes = (Axes) shape;
-    short mad = viewer.getObjectMad(StateManager.OBJ_AXIS1);
+    int mad = viewer.getObjectMad(StateManager.OBJ_AXIS1);
     if (mad == 0 || !g3d.checkTranslucent(false))
       return;
     imageFontScaling = viewer.getImageFontScaling();
